2227: Review fixes
